+++ src/main/java/janggi/board/Palaces.java
@@ -1,5 +1,6 @@
-package janggi;
+package janggi.board;
 
+import janggi.player.Team;
 import janggi.coordinate.Position;
 
 import java.util.List;
