+++ src/main/java/janggi/piece/Cannon.java
@@ -1,7 +1,7 @@
 package janggi.piece;
 
-import janggi.Board;
-import janggi.Team;
+import janggi.board.Board;
+import janggi.player.Team;
 import janggi.coordinate.Position;
 import janggi.coordinate.Route;
 import janggi.piece.rule.movement.MovementRule;
