+++ src/main/java/janggi/player/Turn.java
@@ -1,4 +1,4 @@
-package janggi;
+package janggi.player;
 
 public class Turn {
